pull out requestSocketToken
from MiniXHRUpload
